refactor(outpost): drop ganesha driver and its kerberos sectype
